i-vison-backend/requirements.txt
pymilvus
sentence-transformers
transformers
accelerate
fastapi
uvicorn
Pillow
google-genai
dotenv
ollama
